pyproject.toml
[tool.poetry]
name = "metalrag"
version = "0.1.0"
description = "RAG using information fetched from metallum"
authors = ["Hirofumi Tsuda"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-metallum = "^1.1.0"
groq = "^0.13.0"
pydantic = "^2.10.2"
python-dotenv = "^1.0.1"
requests-cache = "^0.5.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
